requirements.txt
pandas
matplotlib
streamlit
request
numpy